dev/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Session==0.8.0
Werkzeug==3.0.1
python-socketio==5.10.0
python-engineio>=4.8.0
eventlet==0.33.3
psycopg2-binary==2.9.9
Pillow==10.0.0
ldap3==2.9.1
requests==2.32.3
